Moved my files to a folder in the root directory per the instructions and reverted my changes to the original source files.
