+++ Challenge Source/Jonathan_Wiley_solution/LinkedListCodingChallenge.h
@@ -1,0 +1,19 @@
+//
+//  LinkedListCodingChallenge.h
+//  LinkedListCodingChallenge
+//
+//  Created by Patrick Goley on 1/30/18.
+//  Copyright © 2018 nashcocoaheads. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for LinkedListCodingChallenge.
+FOUNDATION_EXPORT double LinkedListCodingChallengeVersionNumber;
+
+//! Project version string for LinkedListCodingChallenge.
+FOUNDATION_EXPORT const unsigned char LinkedListCodingChallengeVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <LinkedListCodingChallenge/PublicHeader.h>
+
+
